moving easymock tests to mockito without update
