+++ google-cloud-storage/src/test/java/com/google/cloud/storage/StorageImplTest.java
@@ -427,325 +427,6 @@ public class StorageImplTest {
   }
 
   @Test
-  public void testUpdateBucket() {
-    BucketInfo updatedBucketInfo = BUCKET_INFO1.toBuilder().setIndexPage("some-page").build();
-    EasyMock.expect(storageRpcMock.patch(updatedBucketInfo.toPb(), EMPTY_RPC_OPTIONS))
-        .andReturn(updatedBucketInfo.toPb());
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    Bucket bucket = storage.update(updatedBucketInfo);
-    assertEquals(new Bucket(storage, new BucketInfo.BuilderImpl(updatedBucketInfo)), bucket);
-  }
-
-  @Test
-  public void testUpdateBucketWithOptions() {
-    BucketInfo updatedBucketInfo = BUCKET_INFO1.toBuilder().setIndexPage("some-page").build();
-    EasyMock.expect(storageRpcMock.patch(updatedBucketInfo.toPb(), BUCKET_TARGET_OPTIONS))
-        .andReturn(updatedBucketInfo.toPb());
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    Bucket bucket =
-        storage.update(
-            updatedBucketInfo, BUCKET_TARGET_METAGENERATION, BUCKET_TARGET_PREDEFINED_ACL);
-    assertEquals(new Bucket(storage, new BucketInfo.BuilderImpl(updatedBucketInfo)), bucket);
-  }
-
-  @Test
-  public void testUpdateBlob() {
-    BlobInfo updatedBlobInfo = BLOB_INFO1.toBuilder().setContentType("some-content-type").build();
-    EasyMock.expect(storageRpcMock.patch(updatedBlobInfo.toPb(), EMPTY_RPC_OPTIONS))
-        .andReturn(updatedBlobInfo.toPb());
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    Blob blob = storage.update(updatedBlobInfo);
-    assertEquals(new Blob(storage, new BlobInfo.BuilderImpl(updatedBlobInfo)), blob);
-  }
-
-  @Test
-  public void testUpdateBlobWithOptions() {
-    BlobInfo updatedBlobInfo = BLOB_INFO1.toBuilder().setContentType("some-content-type").build();
-    EasyMock.expect(storageRpcMock.patch(updatedBlobInfo.toPb(), BLOB_TARGET_OPTIONS_UPDATE))
-        .andReturn(updatedBlobInfo.toPb());
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    Blob blob =
-        storage.update(updatedBlobInfo, BLOB_TARGET_METAGENERATION, BLOB_TARGET_PREDEFINED_ACL);
-    assertEquals(new Blob(storage, new BlobInfo.BuilderImpl(updatedBlobInfo)), blob);
-  }
-
-  @Test
-  public void testDeleteBucket() {
-    EasyMock.expect(storageRpcMock.delete(BucketInfo.of(BUCKET_NAME1).toPb(), EMPTY_RPC_OPTIONS))
-        .andReturn(true);
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    assertTrue(storage.delete(BUCKET_NAME1));
-  }
-
-  @Test
-  public void testDeleteBucketWithOptions() {
-    EasyMock.expect(
-            storageRpcMock.delete(BucketInfo.of(BUCKET_NAME1).toPb(), BUCKET_SOURCE_OPTIONS))
-        .andReturn(true);
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    assertTrue(storage.delete(BUCKET_NAME1, BUCKET_SOURCE_METAGENERATION));
-  }
-
-  @Test
-  public void testDeleteBlob() {
-    EasyMock.expect(
-            storageRpcMock.delete(BlobId.of(BUCKET_NAME1, BLOB_NAME1).toPb(), EMPTY_RPC_OPTIONS))
-        .andReturn(true);
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    assertTrue(storage.delete(BUCKET_NAME1, BLOB_NAME1));
-  }
-
-  @Test
-  public void testDeleteBlobWithOptions() {
-    EasyMock.expect(
-            storageRpcMock.delete(BlobId.of(BUCKET_NAME1, BLOB_NAME1).toPb(), BLOB_SOURCE_OPTIONS))
-        .andReturn(true);
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    assertTrue(
-        storage.delete(
-            BUCKET_NAME1, BLOB_NAME1, BLOB_SOURCE_GENERATION, BLOB_SOURCE_METAGENERATION));
-  }
-
-  @Test
-  public void testDeleteBlobWithOptionsFromBlobId() {
-    EasyMock.expect(storageRpcMock.delete(BLOB_INFO1.getBlobId().toPb(), BLOB_SOURCE_OPTIONS))
-        .andReturn(true);
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    assertTrue(
-        storage.delete(
-            BLOB_INFO1.getBlobId(),
-            BLOB_SOURCE_GENERATION_FROM_BLOB_ID,
-            BLOB_SOURCE_METAGENERATION));
-  }
-
-  @Test
-  public void testCompose() {
-    Storage.ComposeRequest req =
-        Storage.ComposeRequest.newBuilder()
-            .addSource(BLOB_NAME2, BLOB_NAME3)
-            .setTarget(BLOB_INFO1)
-            .build();
-    EasyMock.expect(
-            storageRpcMock.compose(
-                ImmutableList.of(BLOB_INFO2.toPb(), BLOB_INFO3.toPb()),
-                BLOB_INFO1.toPb(),
-                EMPTY_RPC_OPTIONS))
-        .andReturn(BLOB_INFO1.toPb());
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    Blob blob = storage.compose(req);
-    assertEquals(expectedBlob1, blob);
-  }
-
-  @Test
-  public void testComposeWithOptions() {
-    Storage.ComposeRequest req =
-        Storage.ComposeRequest.newBuilder()
-            .addSource(BLOB_NAME2, BLOB_NAME3)
-            .setTarget(BLOB_INFO1)
-            .setTargetOptions(BLOB_TARGET_GENERATION, BLOB_TARGET_METAGENERATION)
-            .build();
-    EasyMock.expect(
-            storageRpcMock.compose(
-                ImmutableList.of(BLOB_INFO2.toPb(), BLOB_INFO3.toPb()),
-                BLOB_INFO1.toPb(),
-                BLOB_TARGET_OPTIONS_COMPOSE))
-        .andReturn(BLOB_INFO1.toPb());
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    Blob blob = storage.compose(req);
-    assertEquals(expectedBlob1, blob);
-  }
-
-  @Test
-  public void testCopy() {
-    CopyRequest request = Storage.CopyRequest.of(BLOB_INFO1.getBlobId(), BLOB_INFO2.getBlobId());
-    StorageRpc.RewriteRequest rpcRequest =
-        new StorageRpc.RewriteRequest(
-            request.getSource().toPb(),
-            EMPTY_RPC_OPTIONS,
-            false,
-            BLOB_INFO2.toPb(),
-            EMPTY_RPC_OPTIONS,
-            null);
-    StorageRpc.RewriteResponse rpcResponse =
-        new StorageRpc.RewriteResponse(rpcRequest, null, 42L, false, "token", 21L);
-    EasyMock.expect(storageRpcMock.openRewrite(rpcRequest)).andReturn(rpcResponse);
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    CopyWriter writer = storage.copy(request);
-    assertEquals(42L, writer.getBlobSize());
-    assertEquals(21L, writer.getTotalBytesCopied());
-    assertTrue(!writer.isDone());
-  }
-
-  @Test
-  public void testCopyWithOptions() {
-    CopyRequest request =
-        Storage.CopyRequest.newBuilder()
-            .setSource(BLOB_INFO2.getBlobId())
-            .setSourceOptions(BLOB_SOURCE_GENERATION, BLOB_SOURCE_METAGENERATION)
-            .setTarget(BLOB_INFO1, BLOB_TARGET_GENERATION, BLOB_TARGET_METAGENERATION)
-            .build();
-    StorageRpc.RewriteRequest rpcRequest =
-        new StorageRpc.RewriteRequest(
-            request.getSource().toPb(),
-            BLOB_SOURCE_OPTIONS_COPY,
-            true,
-            request.getTarget().toPb(),
-            BLOB_TARGET_OPTIONS_COMPOSE,
-            null);
-    StorageRpc.RewriteResponse rpcResponse =
-        new StorageRpc.RewriteResponse(rpcRequest, null, 42L, false, "token", 21L);
-    EasyMock.expect(storageRpcMock.openRewrite(rpcRequest)).andReturn(rpcResponse);
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    CopyWriter writer = storage.copy(request);
-    assertEquals(42L, writer.getBlobSize());
-    assertEquals(21L, writer.getTotalBytesCopied());
-    assertTrue(!writer.isDone());
-  }
-
-  @Test
-  public void testCopyWithEncryptionKey() {
-    CopyRequest request =
-        Storage.CopyRequest.newBuilder()
-            .setSource(BLOB_INFO2.getBlobId())
-            .setSourceOptions(BlobSourceOption.decryptionKey(KEY))
-            .setTarget(BLOB_INFO1, BlobTargetOption.encryptionKey(BASE64_KEY))
-            .build();
-    StorageRpc.RewriteRequest rpcRequest =
-        new StorageRpc.RewriteRequest(
-            request.getSource().toPb(),
-            ENCRYPTION_KEY_OPTIONS,
-            true,
-            request.getTarget().toPb(),
-            ENCRYPTION_KEY_OPTIONS,
-            null);
-    StorageRpc.RewriteResponse rpcResponse =
-        new StorageRpc.RewriteResponse(rpcRequest, null, 42L, false, "token", 21L);
-    EasyMock.expect(storageRpcMock.openRewrite(rpcRequest)).andReturn(rpcResponse).times(2);
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    CopyWriter writer = storage.copy(request);
-    assertEquals(42L, writer.getBlobSize());
-    assertEquals(21L, writer.getTotalBytesCopied());
-    assertTrue(!writer.isDone());
-    request =
-        Storage.CopyRequest.newBuilder()
-            .setSource(BLOB_INFO2.getBlobId())
-            .setSourceOptions(BlobSourceOption.decryptionKey(BASE64_KEY))
-            .setTarget(BLOB_INFO1, BlobTargetOption.encryptionKey(KEY))
-            .build();
-    writer = storage.copy(request);
-    assertEquals(42L, writer.getBlobSize());
-    assertEquals(21L, writer.getTotalBytesCopied());
-    assertTrue(!writer.isDone());
-  }
-
-  @Test
-  public void testCopyFromEncryptionKeyToKmsKeyName() {
-    CopyRequest request =
-        Storage.CopyRequest.newBuilder()
-            .setSource(BLOB_INFO2.getBlobId())
-            .setSourceOptions(BlobSourceOption.decryptionKey(KEY))
-            .setTarget(BLOB_INFO1, BlobTargetOption.kmsKeyName(KMS_KEY_NAME))
-            .build();
-    StorageRpc.RewriteRequest rpcRequest =
-        new StorageRpc.RewriteRequest(
-            request.getSource().toPb(),
-            ENCRYPTION_KEY_OPTIONS,
-            true,
-            request.getTarget().toPb(),
-            KMS_KEY_NAME_OPTIONS,
-            null);
-    StorageRpc.RewriteResponse rpcResponse =
-        new StorageRpc.RewriteResponse(rpcRequest, null, 42L, false, "token", 21L);
-    EasyMock.expect(storageRpcMock.openRewrite(rpcRequest)).andReturn(rpcResponse).times(2);
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    CopyWriter writer = storage.copy(request);
-    assertEquals(42L, writer.getBlobSize());
-    assertEquals(21L, writer.getTotalBytesCopied());
-    assertTrue(!writer.isDone());
-    request =
-        Storage.CopyRequest.newBuilder()
-            .setSource(BLOB_INFO2.getBlobId())
-            .setSourceOptions(BlobSourceOption.decryptionKey(BASE64_KEY))
-            .setTarget(BLOB_INFO1, BlobTargetOption.kmsKeyName(KMS_KEY_NAME))
-            .build();
-    writer = storage.copy(request);
-    assertEquals(42L, writer.getBlobSize());
-    assertEquals(21L, writer.getTotalBytesCopied());
-    assertTrue(!writer.isDone());
-  }
-
-  @Test
-  public void testCopyWithOptionsFromBlobId() {
-    CopyRequest request =
-        Storage.CopyRequest.newBuilder()
-            .setSource(BLOB_INFO1.getBlobId())
-            .setSourceOptions(BLOB_SOURCE_GENERATION_FROM_BLOB_ID, BLOB_SOURCE_METAGENERATION)
-            .setTarget(BLOB_INFO1, BLOB_TARGET_GENERATION, BLOB_TARGET_METAGENERATION)
-            .build();
-    StorageRpc.RewriteRequest rpcRequest =
-        new StorageRpc.RewriteRequest(
-            request.getSource().toPb(),
-            BLOB_SOURCE_OPTIONS_COPY,
-            true,
-            request.getTarget().toPb(),
-            BLOB_TARGET_OPTIONS_COMPOSE,
-            null);
-    StorageRpc.RewriteResponse rpcResponse =
-        new StorageRpc.RewriteResponse(rpcRequest, null, 42L, false, "token", 21L);
-    EasyMock.expect(storageRpcMock.openRewrite(rpcRequest)).andReturn(rpcResponse);
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    CopyWriter writer = storage.copy(request);
-    assertEquals(42L, writer.getBlobSize());
-    assertEquals(21L, writer.getTotalBytesCopied());
-    assertTrue(!writer.isDone());
-  }
-
-  @Test
-  public void testCopyMultipleRequests() {
-    CopyRequest request = Storage.CopyRequest.of(BLOB_INFO1.getBlobId(), BLOB_INFO2.getBlobId());
-    StorageRpc.RewriteRequest rpcRequest =
-        new StorageRpc.RewriteRequest(
-            request.getSource().toPb(),
-            EMPTY_RPC_OPTIONS,
-            false,
-            BLOB_INFO2.toPb(),
-            EMPTY_RPC_OPTIONS,
-            null);
-    StorageRpc.RewriteResponse rpcResponse1 =
-        new StorageRpc.RewriteResponse(rpcRequest, null, 42L, false, "token", 21L);
-    StorageRpc.RewriteResponse rpcResponse2 =
-        new StorageRpc.RewriteResponse(rpcRequest, BLOB_INFO1.toPb(), 42L, true, "token", 42L);
-    EasyMock.expect(storageRpcMock.openRewrite(rpcRequest)).andReturn(rpcResponse1);
-    EasyMock.expect(storageRpcMock.continueRewrite(rpcResponse1)).andReturn(rpcResponse2);
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    CopyWriter writer = storage.copy(request);
-    assertEquals(42L, writer.getBlobSize());
-    assertEquals(21L, writer.getTotalBytesCopied());
-    assertTrue(!writer.isDone());
-    assertEquals(expectedBlob1, writer.getResult());
-    assertTrue(writer.isDone());
-    assertEquals(42L, writer.getTotalBytesCopied());
-    assertEquals(42L, writer.getBlobSize());
-  }
-
-  @Test
   public void testReadAllBytes() {
     EasyMock.expect(
             storageRpcMock.load(BlobId.of(BUCKET_NAME1, BLOB_NAME1).toPb(), EMPTY_RPC_OPTIONS))
